+++ dataiter/data_frame.py
@@ -617,6 +617,8 @@ class DataFrame(dict):
     def read_csv(cls, fname, encoding="utf_8", header=True, columns=None, sep=","):
         """
         Return a new data frame from CSV file `fname`.
+
+        Will automatically decompress if `fname` ends in ``.bz2|.gz|.xz``.
         """
         import pandas as pd
         data = pd.read_csv(fname,
@@ -634,8 +636,10 @@ class DataFrame(dict):
     def read_json(cls, fname, encoding="utf_8", **kwargs):
         """
         Return a new data frame from JSON file `fname`.
+
+        Will automatically decompress if `fname` ends in ``.bz2|.gz|.xz``.
         """
-        with open(fname, "r", encoding=encoding) as f:
+        with util.xopen(fname, "rt", encoding=encoding) as f:
             return cls.from_json(f.read(), **kwargs)
 
     @classmethod
@@ -654,8 +658,10 @@ class DataFrame(dict):
     def read_pickle(cls, fname):
         """
         Return a new data frame from Pickle file `fname`.
+
+        Will automatically decompress if `fname` ends in ``.bz2|.gz|.xz``.
         """
-        with open(fname, "rb") as f:
+        with util.xopen(fname, "rb") as f:
             return cls(pickle.load(f))
 
     def _reconcile_column(self, column):
@@ -905,6 +911,8 @@ class DataFrame(dict):
     def write_csv(self, fname, encoding="utf_8", header=True, sep=","):
         """
         Write data frame to CSV file `fname`.
+
+        Will automatically compress if `fname` ends in ``.bz2|.gz|.xz``.
         """
         pddf = self.to_pandas()
         util.makedirs_for_file(fname)
@@ -913,6 +921,8 @@ class DataFrame(dict):
     def write_json(self, fname, encoding="utf_8", **kwargs):
         """
         Write data frame to JSON file `fname`.
+
+        Will automatically compress if `fname` ends in ``.bz2|.gz|.xz``.
 
         `kwargs` are passed to ``json.JSONEncoder``.
         """
@@ -929,8 +939,10 @@ class DataFrame(dict):
     def write_pickle(self, fname):
         """
         Write data frame to Pickle file `fname`.
+
+        Will automatically compress if `fname` ends in ``.bz2|.gz|.xz``.
         """
         util.makedirs_for_file(fname)
-        with open(fname, "wb") as f:
+        with util.xopen(fname, "wb") as f:
             out = {k: np.array(v, v.dtype) for k, v in self.items()}
             pickle.dump(out, f, pickle.HIGHEST_PROTOCOL)